+++ RevitAddin.CommandBindingTester/Models/BeforeExecutedCommand.cs
@@ -1,0 +1,10 @@
+﻿namespace RevitAddin.CommandBindingTester.Models
+{
+    public enum BeforeExecutedCommand
+    {
+        None,
+        ShowMessage,
+        ThrowException,
+        Cancel,
+    }
+}