+++ RevitAddin.CommandBindingTester/Models/CanExecuteCommand.cs
@@ -1,0 +1,11 @@
+﻿namespace RevitAddin.CommandBindingTester.Models
+{
+    public enum CanExecuteCommand
+    {
+        None,
+        Always,
+        Never,
+        WhenDocument,
+        WhenFamily,
+    }
+}