+++ RevitAddin.CommandBindingTester/Models/ExecutedCommand.cs
@@ -1,0 +1,10 @@
+﻿namespace RevitAddin.CommandBindingTester.Models
+{
+    public enum ExecutedCommand
+    {
+        None,
+        ShowMessage,
+        ThrowException,
+        TransactionAuthor,
+    }
+}