+++ RevitAddin.CommandBindingTester/ViewModels/CreateBindingViewModel.cs
@@ -1,4 +1,5 @@
-﻿using RevitAddin.CommandBindingTester.Models;
+﻿using Autodesk.Revit.UI;
+using RevitAddin.CommandBindingTester.Models;
 using RevitAddin.CommandBindingTester.Views;
 using ricaun.Revit.Mvvm;
 using ricaun.Revit.UI;
@@ -13,6 +14,8 @@ namespace RevitAddin.CommandBindingTester.ViewModels
     [PropertyChanged.AddINotifyPropertyChangedInterface]
     public class CreateBindingViewModel : ObservableObject
     {
+        public static CreateBindingViewModel Instance { get; } = new CreateBindingViewModel();
+
         #region Public Properties
         public ObservableCollection<CreateBindingModel> Models { get; } = new();
         public CreateBindingModel SelectedItem { get; set; }
@@ -22,12 +25,6 @@ namespace RevitAddin.CommandBindingTester.ViewModels
         private bool CanDeleteSelected()
         {
             return SelectedItem is not null;
-            //if (CurrentCell is DataGridCellInfo data)
-            //{
-            //    if (data.Item is CreateBindingModel)
-            //        return true;
-            //}
-            //return false;
         }
 
         private void DeleteSelected()
@@ -42,12 +39,32 @@ namespace RevitAddin.CommandBindingTester.ViewModels
         #region Constructor
         public CreateBindingViewModel()
         {
-
+            Models.Add(new CreateBindingModel()
+            {
+                PostableCommand = PostableCommand.ThinLines,
+                CanExecute = CanExecuteCommand.None,
+                BeforeExecuted = BeforeExecutedCommand.None,
+                Executed = ExecutedCommand.ShowMessage,
+            });
+            Models.Add(new CreateBindingModel()
+            {
+                PostableCommand = PostableCommand.ThinLines,
+                CanExecute = CanExecuteCommand.WhenFamily,
+                BeforeExecuted = BeforeExecutedCommand.None,
+                Executed = ExecutedCommand.TransactionAuthor,
+            });
+            Models.Add(new CreateBindingModel()
+            {
+                PostableCommand = PostableCommand.ThinLines,
+                CanExecute = CanExecuteCommand.WhenDocument,
+                BeforeExecuted = BeforeExecutedCommand.None,
+                Executed = ExecutedCommand.TransactionAuthor,
+            });
         }
         #endregion
 
         #region View / Window
-        public string Title { get; set; } = "CreateBindingViewModel";
+        public string Title { get; set; } = "RevitCommand Binding Editor";
         public CreateBindingView Window { get; private set; }
         public void Show()
         {